+++ src/main/java/portfolio/Portfoliomonthly.java
@@ -1,11 +1,13 @@
 package portfolio;
 
+import java.math.BigDecimal;
+
 public class Portfoliomonthly {
   private Integer id;
   private Integer totaltouristcount;
   private Integer armtouristcount;
   private Integer othertouristcount;
-  private Double finances;
+  private BigDecimal finances;
   private Integer portfolioid;
   private Integer monthId;
 
@@ -41,11 +43,11 @@ public class Portfoliomonthly {
     this.othertouristcount = othertouristcount;
   }
 
-  public Double getFinances() {
+  public BigDecimal getFinances() {
     return finances;
   }
 
-  public void setFinances(Double finances) {
+  public void setFinances(BigDecimal finances) {
     this.finances = finances;
   }
